fix(protocol): header v2 — timestamp ÷1000, sequence u16, context_version u24
Fix 1: Timestamp stored as Unix seconds (÷1000) instead of truncated
milliseconds — eliminates silent 49-day wrap bug.
encode_raw() now uses context.version() instead of hardcoded 0.

Fix 2: MessageHeader::sequence reduced from u32 (4B) to u16 (2B).
Wraps every 65 536 frames — standard rollover pattern.

Fix 3: MessageHeader::context_version serialized as u24 (3B on wire,
u32 in Rust). Max 16 777 215 versions before rollover.

MessageHeader::SIZE: 13 → 10 (saves 3 bytes per frame).

Includes 8 regression tests in tests/protocol_header_v2.rs.
Version bump deferred to after name_id fix.

+++ src/channel.rs
@@ -274,7 +274,7 @@ mod tests {
     use super::*;
     use crate::protocol::{MessageHeader, MessageType, Priority};
 
-    fn make_test_message(seq: u32) -> EncodedMessage {
+    fn make_test_message(seq: u16) -> EncodedMessage {
         EncodedMessage::new(
             MessageHeader {
                 version: 1,
